requirements.txt
flask
notebook
paho-mqtt
pymongo
haversine
numpy
elasticsearch
pandas
matplotlib
seaborn
rankFM
sklearn
openpyxl
